fix(server-go): preserve Node numeric coercion

+++ apps/server-go/internal/content/model.go
@@ -42,7 +42,7 @@ type Author struct {
 }
 
 type ListQuery struct {
-	Page     int
+	Page     float64
 	Limit    int
 	Category string
 	Search   string
